Update requirements and refactor gym_api imports
- Added `python-dotenv` version 1.1.0 to requirements.txt.
- Refactored import paths for `GymApi` from `evo_client.api` to `evo_client.services` in multiple example scripts and test files.
- Introduced a new `gym_api.py` service file to encapsulate Gym API client functionality.
