+++ examples/fetch_members_files.py
@@ -4,7 +4,7 @@ from typing import List
 
 from evo_client.core.configuration import Configuration
 from evo_client.core.api_client import ApiClient
-from evo_client.api.gym_api import GymApi
+from evo_client.services.gym_api import GymApi
 from evo_client.models.gym_model import MembershipStatus, MembersFiles
