+++ examples/fetch_operating_data.py
@@ -5,7 +5,7 @@ import asyncio
 
 from evo_client.core.configuration import Configuration
 from evo_client.core.api_client import ApiClient
-from evo_client.api.gym_api import GymApi
+from evo_client.services.gym_api import GymApi
 from evo_client.models.gym_model import GymOperatingData
 from evo_client.services.data_fetchers import BranchApiClientManager
