+++ examples/gym_api_example.py
@@ -3,7 +3,7 @@ from decimal import Decimal
 from multiprocessing.pool import AsyncResult
 from typing import cast, List
 
-from evo_client.api.gym_api import GymApi
+from evo_client.services.gym_api import GymApi
 from evo_client.models.gym_model import (
     NewSale,
     PaymentMethod,
