+++ examples/webhook_management_example.py
@@ -1,7 +1,7 @@
 from typing import cast, List
 from multiprocessing.pool import AsyncResult
 
-from evo_client.api.gym_api import GymApi
+from evo_client.services.gym_api import GymApi
 from evo_client.models.w12_utils_webhook_view_model import W12UtilsWebhookViewModel
 
 # Initialize the API client
